Ok, head back up to the action method. Hmm, so all we really need to do now is call the cancel() method on {dollar_sign}subscription! But let's get a little bit more organized. Open SubscriptionHelper and add a new method there: public function fullyCancelSubscription() with the Subscription object that should be canceled. Below, really simple, say {dollar_sign}subscription->cancel(). Then, use the Doctrine entity manager to save this to the database. Back in the controller, call this! Above the switch statement, add a {dollar_sign}subscriptionHelper variable set to {dollar_sign}this->get('subscription_helper'). Finally, call {dollar_sign}subscriptionHelper->fullyCancelSubscription({dollar_sign}subscription).
And that is it!

Yep, there's some setup to get the webhook controller started, but now we're in really good shape.

Of course... we have no way to test this... So, ya know, just make sure you do a really good job of coding and hope for the best! No, that's crazy!, I'll show you a few ways to test this next. But also, don't forget to configure your webhook URL in Stripe once you finally deploy this to beta and production.