The reason that webhooks are so hard is... well, they're kind of impossible to test. There are certain things on Stripe - like my card being declined during a renewal, which are really hard to simulate. And even if we could, Stripe can't send a webhook to my local computer. Well actually, that last part isn't entirely true, but more on that later.
So let's look at a few strategies for making sure that your webhooks are air tight. I do not want to mess something up with these.

Automated Test

The first strategy - and the one we use here on KnpU - is to create an automated test that sends a fake webhook to the URL.

To start, let's install PHPUnit into the project
composer require phpunit/phpunit --dev