To send the request, create a  variable set to  want, you can also use something different, like Guzzle. It doesn't really matter because - one way or another - you just need to make an HTTP request to the endpoint.
Now for the magic: call {dollar_sign}client->request() and pass it a bunch of arguments: POST for the HTTP method, /webhooks/stripe, then a few empty arrays for parameters, files and server. Finally, for the {dollar_sign}content argument - the body of the request - pass it {dollar_sign}eventJson.

And because things almost never work for me on the first try... and because I know this won't work yet, let's dump({dollar_sign}client->getResponse()->getContent()) to see what happened in case there's an error. Also add a sanity check, {dollar_sign}this->assertEquals() that 200 matches {dollar_sign}client->getResponse()->getStatusCode()