Let's run the test. Copy its method name, then open your terminal. It looks like PHPUnit installed just fine. So, run:
./vendor/bin/phpunit --filter testStripeCustomerSubscriptionDeleted
Oh no! It blew up! Hmmm:

Unknown database 'stripe_recording_test'
Ah, my bad!

I setup our project to use a different database for testing... and I forgot to create it! Do that with:

./bin/console doctrine:database:create --env=test
And to create the tables, run:

./bin/console doctrine:schema:create --env=test
Try the test again:

./vendor/bin/phpunit --filter testStripeCustomerSubscriptionDeleted
Another error! Scroll to the top! The webhook returned a 500 error. And if you look closely at the dumped response HTML, you can see the reason:

No such event: evt_00000000000000
Ah, the id of the fake event that we're sending is evt_00000000000000. That's not a real event in Stripe, and so when the WebhookController reads this and uses Stripe's API to fetch this event, it's not there:

 61 lines  src/AppBundle/Controller/WebhookController.php
... lines 1 - 8
class WebhookController extends BaseController
{
... lines 11 - 13
  public function stripeWebhookAction(Request {dollar_sign}request)
  {
... lines 16 - 22
    {dollar_sign}stripeEvent = {dollar_sign}this->get('stripe_client')
      ->findEvent({dollar_sign}eventId);
... lines 25 - 39
  }
... lines 41 - 60
}
It's kind of funny: we added this API lookup to prevent a third-party from sending fake events... and now it's stopping us from doing exactly that. Dang!