Hmm, how to fix this? In the real world, we do want to use stripe's API to fetch the Event object. But in the test environment, this would all work if our code would simply use the JSON we're sending it as the event, and skip the lookup.
Let's do it! We'll set a special configuration variable in the test environment only, then use that to change our logic in the controller.

Open app/config/config.yml and add a new parameter: verify_stripe_event set to true.
Copy that, and open config_test.yml. Add a parameters key, paste this parameter, but override it to be false.

Now, in WebhookController, we just need an if statement: if {dollar_sign}this->getParameter('verify_stripe_event') is true, then keep the normal behavior. Otherwise, set {dollar_sign}stripeEvent to json_decode({dollar_sign}request->getContent()).

OK, this is not technically perfect: the first  is a \Stripe\Event object, and the second will be an instance of stdClass. But, since you fetch data off both the same way, it should work.

Let's see if does! Try the test again:

./vendor/bin/phpunit --filter testStripeCustomerSubscriptionDeleted
This time, no errors! And the dumped response content looks perfect: event handled.