Refreshing the Data after the Test
But, the test didn't pass:

Failed asserting that true is false on line 43
It looks like our webhook is not working, because the subscription is still active. But actually, that's not true: Doctrine is tricking us! In reality, the database has been updated to show that the Subscription is canceled, but this Subscription object is out-of-date. Query for a fresh one with  = {dollar_sign}this->em - I set the EntityManager on that property in setup() - then ->getRepository('AppBundle:Subscription') with find({dollar_sign}subscription->getId()).

This subscription will have fresh data.

Try the test!

./vendor/bin/phpunit --filter testStripeCustomerSubscriptionDeleted
And we are green!

I know, that was kind of hard! But if you want to have automated webhook tests, this is the way to do it. To make matters worse, for other webhooks, you may need to fake additional API calls that you're making to Stripe.

But, there are also a couple of other, manual, but easy ways to test. Let's check 'em out!