Updating the Subscription in the Database
Finally, let's update the Subscription in our database by using the data on the Stripe subscription object. As usual, we'll add this logic to SubscriptionHelper so we can reuse it later.

Add a new public function called handleSubscriptionPaid() that has two arguments: the Subscription object that just got paid and the related \Stripe\Subscription object that holds the updated details.

Then, we need to read the current_period_end field. But wait! We totally did this earlier in addSubscriptionToUser(). Steal that line! Paste it here, but rename the variable to {dollar_sign}newPeriodEnd.
Now, set this billingPeriodEndsAt field via {dollar_sign}subscription->setBillingPeriodEndsAt().
But wait! Where's my auto-completion! Oh, that method doesn't exist yet. In Subscription, I'll use the 'Code'->'Generate' shortcut to select 'Setters' and generate this setter.
Persist and flush the Subscription changes to the database.
Back in your controller, call this: {dollar_sign}subscriptionHelper->handleSubscriptionPaid() and pass it {dollar_sign}subscription and {dollar_sign}stripeSubscription.

I won't test this - let's call that homework for you - but now whenever a subscription is renewed, the  will be updated.