But there's just one other small important thing you'll want to do each time a subscription is renewed: send the user an email! Ok, we're not actually going to code up the email-sending logic now - but it would live right here in SubscriptionHelper.
But wait! There is one gotcha: the invoice.payment_succeeded webhook will be triggered when a subscription is renewed... but it will also be triggered at the moment that the user originally buys their subscription. So if you send your user an email that says: 'thanks for renewing your subscription', then any new users will be pretty confused.

To fix this, add a new {dollar_sign}isRenewal variable set to {dollar_sign}newPeriodEnd > {dollar_sign}subscription->getBillingPeriodEndsAt().

If this is a new subscription, then it was completed about 2 seconds ago, and we would have already set the billingPeriodEndsAt to the correct date. When the webhook fires, the dates will already match. But if this is a renewal, the billingPeriodEndsAt in the database will be for last month, and {dollar_sign}newPeriodEnd will be for next month.

In other words, you can use the {dollar_sign}isRenewal flag to send the right type of email.