Let's prevent that. And it's simple: create a database table that records all the event ID's we've handled. Then, query that table before processing a webhook to make sure we haven't seen it before.
In the AppBundle/Entity directory, create a new PHP Class called StripeEventLog.
Give it a few properties: {dollar_sign}id, {dollar_sign}stripeEventId and a {dollar_sign}handledAt date field.
Since this project uses Doctrine, I'll add a special use statement on top and then add some annotations, so that this new class will become a new table in the database. Use the 'Code'->'Generate' menu, or Command + N on a Mac and select 'ORM Class'.
Repeat that and select 'ORM Annotations'. Choose all the fields.
Update stripeEventId to be a string field - that'll translate to a varchar in MySQL.
To set the properties, create a new __construct() method with a  argument. Inside, set that on the property and also set {dollar_sign}this->handledAt to a new \DateTime() to set this field to 'right now'