To make sure we didn't mess anything up, open WebhookControllerTest and copy our test method. Run that:
./vendor/bin/phpunit --filter testStripeCustomerSubscriptionDeleted
Bah! Of course... it failed for a silly reason: I need to update my test database - to add the new table. A shortcut to do that is:

./bin/console doctrine:schema:update --force --env=test
Try the test now:

./vendor/bin/phpunit --filter testStripeCustomerSubscriptionDeleted
It works! So hey, run it again!

./vendor/bin/phpunit --filter testStripeCustomerSubscriptionDeleted
It fails?!

Failed to assert that true is false.
Well, that's not clear, but I know what the problem is: every event in the test has the same event ID.
So when you run the test the second time, this already exists in the StripeEventLog table and the webhook is skipped. Well hey, at least we know the replay attack system is working.

To fix this, we need to set a little bit of randomness to the event ID by adding a %s at the end and adding an mt_rand() to the sprintf().

Now, every event ID will be unique. Try the test again:

./vendor/bin/phpunit --filter testStripeCustomerSubscriptionDeleted
Green and happy!

Ok, enough webhooks. Let's do something fun, like making it possible for a user to upgrade from one subscription to another.