Now, step two: add an upgrade button that mentions the plan they could switch to. Since we only have 2 plans, it's pretty simple: if they're on the Farmer Brent, we want to allow them to upgrade to the New Zealander. And if they're on the New Zealander, we should let them downgrade to the Farmer Brent.
To find the other plan, open SubscriptionHelper and add a new public function called findPlanToChangeTo with a {dollar_sign}currentPlanId argument.

I'll paste in the logic: it's kind of silly, but it gets the job done. I'm using str_replace instead of something simpler, because in a few minutes, we're going to add yearly plans, and I still want this function to... um... function.

Back to the controller! Add another variable:  = null. Then, {dollar_sign}otherPlan = {dollar_sign}this->get('subscription_helper')->findPlanToChangeTo() and pass it {dollar_sign}currentPlan->getPlanId(). Pass this into the template as an otherPlan variable.

There, after the 'Active' label, add a button with some classes: a few for styling, one to float right and one - js-change-plan-button - that we'll use in a minute via JavaScript. Make the text: 'Change to' and then otherPlan.name.
Oh, and add one more attribute: data-plan-name and print otherPlan.name. We'll read that attribute in JavaScript.