Honestly, upgrade and downgrading a plan would be really easy, except that we need to calculate how much we should charge the user and tell them so they can confirm.
Getting this right takes some work, but the result is going to be gorgeous, I promise. Here's the plan: as soon as this screen loads, we'll make an AJAX call back to the server. The server will calculate how much to charge the customer for the upgrade, and send that back so we can show it.

In ProfileController, add a new public function called previewPlanChangeAction(). Set the URL to /profile/plan/change/preview/{planId} and give it a name: account_preview_plan_change.

Use the {dollar_sign}planId in the route to load a  object with {dollar_sign}this->get('subscription_helper') and then call findPlan() with {dollar_sign}planId