Back in the controller, call this with {dollar_sign}stripeSubscription = {dollar_sign}this->get('stripe_client') then ->changePlan({dollar_sign}this->getUser(), {dollar_sign}plan).
Upgrading the Plan in our Database

Ok, the plan is upgraded! Well, in Stripe. But we also need to update the subscription row in our database.

When a user buys a new subscription, we call a method on SubscriptionHelper called addSubscriptionToUser(). We pass it the new \Stripe\Subscription and the User. Then it guarantees that the user has a subscription row in the table with the correct data, like the plan id, subscription id, and {dollar_sign}periodEnd date.
Now, the only thing we need to update right now is the plan ID: both the subscription ID and period end haven't changed. But that's ok: we can still safely reuse this method.

In ProfileController, add {dollar_sign}this->get('subscription_helper')->addSubscriptionToUser() passing it {dollar_sign}stripeSubscription and {dollar_sign}this->getUser().

And that's everything. At the bottom... well, we don't really need to return anything to our JSON. So just return a new Response() with null as the content and a 204 status code.
This doesn't do anything special: 204 simply means that the operation was successful, but the server has nothing it wishes to say back.