Executing the Upgrade in the UI
Copy the route name, then head to the template to make this work.

First, find the button, copy the data-preview-url attribute, and paste it. Name the new one data-change-url and update the route name.
Above in the JavaScript, set a new changeUrl variable to {dollar_sign}(this).data('change-url').
Then, scroll down to the bottom: this callback function will be executed if the user clicks the 'Ok' button to confirm the change. Make the AJAX call here: set the url to changeUrl, the method to POST, and attach one more success function.
Inside that, call Sweet Alert to tell the user that the plan was changed! Let's also add some code to reload the page after everything.
OK! Let's do this! Refresh the page! Click to change to the 'New Zealander'. {dollar_sign}99.88 - that looks right, now press 'Ok'. And ... cool! I think it worked! When the page reloads, our plan is the 'New Zealander' and we can downgrade to the 'Farmer Brent'.

In the Stripe dashboard, open payments, click the one for 9.88, and open its Invoice. Oh, it's a thing of beauty: this has the two line items for the change.

If you check out the customer, their top subscription is now to the New Zealander plan.

So we're good. Except for one last edge-case.