This whole mess starts in StripeClient, when we call {dollar_sign}this->createInvoice(), because this might fail.Scroll up to that method. In addition to calling this to upgrade a subscription plan, we also call this at checkout, even if there is no subscription. The problem is that if payment fails on checkout, this invoice will still exist, and Stripe will try again to charge it. Imagine having your card be declined at checkout, only for the vendor to try to charge it again later, without you ever having finished the checkout process!
Here's our rescue plan: if paying the invoice fails, we need to close it. By doing that, Stripe will not try to pay it again.
To do that, surround the pay() line with a try-catch for the \Stripe\Error\Card exception. Here, add {dollar_sign}invoice->close = true and then {dollar_sign}invoice->save(). Then, re-throw the exception. Our checkout logic looks for this exception and uses it to notify the user of the problem.
Next, down in the other function, if we fail to create the invoice, we need to not change the customer's plan in Stripe.

Add a new variable called  set to ->plan->id. Then, surround the createInvoice() call with a try-catch block for the same exception: \Stripe\Error\Card