Here's the goal: on the account page, next to the 'Next Billing at' text, I want to add a link that says 'bill yearly' or 'bill monthly'. When you click this, it should follow the same workflow we just built for upgrading a plan: it should show the cost, then make the change.
In ProfileController::accountAction(), add yet another variable here called {dollar_sign}otherDurationPlan.
This will eventually be the SubscriptionPlan object for the other duration of the current plan. So if I have the monthly Farmer Brent, this will be set to the yearly Farmer Brent plan.

To find that plan, open SubscriptionHelper and add a new function called findPlanForOtherDuration() with a {dollar_sign}currentPlanId argument. I'll paste in some silly code here. This relies on our naming conventions to switch between monthly and yearly plans.