Back in the controller, copy the {dollar_sign}otherPlan line, paste it, then update the variable to {dollar_sign}otherDurationPlan and the method to findPlanForOtherDuration(). Pass that into the template as a new variable.
In account.html.twig, scroll down to the Upgrade Plan button. Copy that whole thing. Then, keep scrolling to the 'Next Billing at' section. If the user has a subscription, paste the upgrade button. And since this process will be the same as upgrading, we can re-use this exactly. Just change otherPlan to otherDurationPlan... in all 4 places. Update the text to 'Bill' and then otherDurationPlan.duration. So, this will say something like 'Bill yearly'.