Dump the Upcoming Invoice
Before we try this, go back into ProfileController and find previewPlanChangeAction(). The truth is, changing a plan from monthly to yearly should be identical to upgrading a plan. But, it's not quite the same. To help us debug an issue we're about to see, dump the {dollar_sign}stripeInvoice variable.

And now that I've told you it won't work, let's try it out! Refresh the account page. Then click the new 'Bill yearly' link. Ok:

You will be charged {dollar_sign}792.05 immediately
Wait, that doesn't seem right. The yearly plan is {dollar_sign}990 per year. Then, if you subtract approximately 9 from that as a credit, it should be something closer to {dollar_sign}891. Something is not quite right.