Duration Change? Don't Invoice
Because of this behavior difference when the duration changes, we need to fix one other spot: in StripeClient::changePlan(). Right now, we manually create an invoice so that the customer is charged immediately. But... we don't need to do that in this case: Stripe automatically creates and pays an Invoice when the duration changes.

In fact, trying to create an invoice will throw an error! Let's see it. First, update your credit card to one that will work.

Now, change to Bill yearly and confirm. The AJAX call should fail... and it does! Open the profiler for that request and find the Exception:

Nothing to invoice for customer
Obviously, we need to avoid this. In StripeClient, add a new variable: {dollar_sign}currentPeriodStart that's set to {dollar_sign}stripeSubscription->current_period_start.

That's the current period start date before we change the plan.

After we change the plan, if the duration is different, the current period start will have changed. Surround the entire invoicing block with if {dollar_sign}stripeSubscription->current_period_start == {dollar_sign}currentPeriodStart. In other words: only invoice the customer manually if the subscription period hasn't changed. I think we should add a note above this: this can look really confusing!

Take it for a Test Drive

But, now it should work! Reset things by going to the pricing page and buying a brand new monthly subscription. Now, head to your account page and update it to yearly. The amount - {dollar_sign}891 - looks right, so hit OK.

Yes! Plan changed! My option changed to 'Bill monthly' and the 'Next Billing at' date is August 10th - one year from today. We should probably print the year.

In Stripe, under payments, we have one for 91, and the customer is on the Farmer Brent yearly plan.

Winning!